requirements.txt
streamlit
deepchem
rdkit-pypi
scikit-learn
joblib
